pyproject.toml
[tool.poetry]
name = "tg-bot-filter"
version = "0.1.0"
description = ""
authors = ["Moffaty <goga459972@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.18.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.38"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
